+++ src/index.ts
@@ -1,5 +1,5 @@
 import express from 'express'
-import employeeRouter from './student.routes'
+import employeeRouter from './employee.routes'
 const app = express()
 
 app.use(express.json())
